Add the loader functionality and fix typo error in app.js
Add a loader icon in the folder and import it in the
index.html and write logic for it to show only when
the app loads and fetches data from the API initially.
The loader should not show in further fetch requests.

Fix the typo. In infinite scroll feature, "ready" variable
which is set to false should be "readyToLoad".